// apps/billing/barrings-mfe-e2e/src/support/enum/barrings-group-name.enum.ts
export enum BarringsGroupName {
  General = 0,
  Roaming = 1,
  PremiumVoice = 2,
  PremiumSms = 3,
  PayByMobile = 4,
}

export enum ProductStatus {
  ACTIVE = 'ACTIVE',
  SUSPENDED = 'SUSPENDED',
  DISCONNECTED = 'DISCONNECTED',
  ACTIVATION_IN_PROGRESS = 'ACTIVATION_IN_PROGRESS',
  DISCONNECTION_IN_PROGRESS = 'DISCONNECTION_IN_PROGRESS',
}